added plotting util + changed validity
